lighting/base/requirements.txt
paho-mqtt==1.5.0
dataclasses-json==0.5.2
